Add e2e runtime test suites for Java/MySQL and Java/SQLite
- Introduced test projects to validate generated query execution for MySQL and SQLite targets in Java.
- MySQL suite uses isolated test databases for full encapsulation (`docker-compose` with per-test database setup/teardown).
- SQLite suite uses in-memory databases for lightweight, fast tests.
- 42 tests per suite, covering a wide range of query types including aggregates, joins, CTEs, subqueries, and parameterized queries.
- Added `Makefile` targets for test generation, execution, and cleanup.
- Extended `.gitignore` to exclude build artifacts and generated files.
- All tests passed in both suites.

+++ src/frontend/mysql/query.rs
@@ -18,9 +18,9 @@ pub(crate) fn parse_queries(sql: &str, schema: &Schema) -> anyhow::Result<Vec<Qu
         schema,
         &ResolverConfig {
             sum_integer_type: SqlType::Decimal,
-            // MySQL: SUM(bigint) → decimal, AVG(integer/bigint) → double
+            // MySQL: SUM/AVG of integer columns → DECIMAL on the wire (not DOUBLE)
             sum_bigint_type: SqlType::Decimal,
-            avg_integer_type: SqlType::Double,
+            avg_integer_type: SqlType::Decimal,
             typemap: crate::frontend::mysql::typemap::map,
         },
     )
